Add Roleplay option to Taboo section
* Adding new kinks that were previously missing
* Adding "Roleplay" option to Taboo section
* Minor rearrangement of existing kinks

There seemed to be a lack of a "roleplay only" option for some of the kinks. As an example, someone might be turned on by the idea of exhibitionism, but would never actually participate in it. To handle this, I've added a "Roleplay" option to Taboo and moved some of the existing kinks also under Taboo like Impregnation and Cuckolding. 

I've also expanded on the list of kinks to include things like Bimbofication, Free use, etc.

Lastly, I've updated the wording of "MtF Trans" and "FtM Trans" to just be "trans female" and "trans male" respectively.